requirements.txt
streamlit>=1.36.0
PyPDF2>=3.0.1
python-dotenv>=1.0.1
cohere>=5.5.8
openai>=1.40.0
faiss-cpu>=1.8.0
numpy>=1.26.4
pandas>=2.2.2
supabase>=1.0.0
matplotlib>=3.8.0
streamlit-chat>=0.0.2